KRACOEUS-6956 : repack propdev clean up proposaldevelopment authorizer fix upon comments
