+++ coeus-code/src/main/java/org/kuali/coeus/propdev/impl/docperm/ModifyProposalPermissionsAuthorizer.java
@@ -15,7 +15,7 @@
  */
 package org.kuali.coeus.propdev.impl.docperm;
 
-import org.kuali.coeus.propdev.impl.core.ProposalAuthorizer;
+import org.kuali.coeus.propdev.impl.auth.task.ProposalAuthorizer;
 import org.kuali.coeus.propdev.impl.core.ProposalDevelopmentDocument;
 import org.kuali.coeus.sys.framework.workflow.KcWorkflowService;
 import org.kuali.kra.infrastructure.PermissionConstants;
