+++ coeus-code/src/main/java/org/kuali/coeus/propdev/impl/editable/AlterProposalDataAuthorizer.java
@@ -20,7 +20,7 @@ import org.kuali.coeus.sys.framework.workflow.KcWorkflowService;
 import org.kuali.kra.infrastructure.PermissionConstants;
 import org.kuali.coeus.propdev.impl.auth.task.ProposalTask;
 import org.kuali.coeus.propdev.impl.hierarchy.ProposalHierarchyException;
-import org.kuali.coeus.propdev.impl.core.ProposalAuthorizer;
+import org.kuali.coeus.propdev.impl.auth.task.ProposalAuthorizer;
 import org.kuali.kra.proposaldevelopment.hierarchy.service.ProposalHierarchyService;
 
 /**
