+++ coeus-code/src/main/resources/org/kuali/kra/datadictionary/docs/SponsorMaintenanceDocument.xml
@@ -16,7 +16,7 @@
       </list>
     </property>
     <property name="documentTypeName" value="SponsorMaintenanceDocument"/>
-    <property name="documentAuthorizerClass" value="org.kuali.coeus.propdev.impl.docperm.SponsorDocumentAuthorizer"/>
+    <property name="documentAuthorizerClass" value="org.kuali.coeus.common.impl.sponsor.SponsorDocumentAuthorizer"/>
     <property name="webScriptFiles">
         <list>
             <value>../dwr/interface/StateService.js</value>
